[AN] refactor: 로그인 로직 UseCase 적용 #718
- [AN] refactor: 로그인 관련 로직 UseCase로 분리
- [AN] refactor: 비로그인 둘러보기 시 불필요한 데이터 삭제 로직 제거
- [AN] chore: GetRecentHearitUseCase 패키지 이동
